Updated header and footer
